+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
-- Nothing should go in this section, please add to the latest unreleased version
-  (and update the corresponding date), or add a new version.
+### Added
+- New API endpoint allowing for easy configuration/persistence of authenticator
+  webservices
+  [#2522](https://github.com/cyberark/conjur/pull/2522)
 
 ## [1.19.6] - 2023-07-05